+++ react/components/LoginComponent.js
@@ -1,6 +1,5 @@
 import React, { Component, Fragment, Suspense } from 'react'
 import {
-  Link,
   withRuntimeContext,
   ExtensionPoint,
   useChildBlock,
@@ -55,26 +54,19 @@ class LoginComponent extends Component {
       iconLabel: iconLabelProfile,
       labelClasses,
       intl,
-      renderIconAsLink,
+      loginButtonAsLink,
       onProfileIconClick,
       sessionProfile,
       showIconProfile,
-      runtime: { history },
+      runtime: { history, navigate },
     } = this.props
 
     const pathname = history && history.location && history.location.pathname
 
     const iconClasses = 'flex items-center'
     const iconLabel = iconLabelProfile || translate('store/login.signIn', intl)
-    const iconContent = (
+    const buttonContent = (
       <Fragment>
-        {showIconProfile && renderIconAsLink && (
-          <ProfileIcon
-            iconSize={iconSize}
-            labelClasses={labelClasses}
-            iconClasses={iconClasses}
-          />
-        )}
         {sessionProfile ? (
           <span
             className={`${styles.profile} t-action--small order-1 pl4 ${labelClasses} dn db-l`}
@@ -94,18 +86,25 @@ class LoginComponent extends Component {
       </Fragment>
     )
 
-    if (renderIconAsLink) {
+    if (loginButtonAsLink) {
       const linkTo = sessionProfile ? 'store.account' : 'store.login'
       const returnUrl =
         !sessionProfile && `returnUrl=${encodeURIComponent(pathname)}`
       return (
-        <Link
-          page={linkTo}
-          query={returnUrl}
-          className={`${styles.buttonLink} h1 w2 tc flex items-center w-100-s h-100-s pa4-s`}
-        >
-          {iconContent}
-        </Link>
+        <div className={styles.buttonLink}>
+          <ButtonWithIcon
+            variation="tertiary"
+            icon={
+              showIconProfile && (
+                <ProfileIcon iconSize={iconSize} labelClasses={labelClasses} iconClasses={iconClasses} />
+              )
+            }
+            iconPosition={showIconProfile ? 'left' : 'right'}
+            onClick={() => navigate({ page: linkTo, query: returnUrl })}
+          >
+            {buttonContent}
+          </ButtonWithIcon>
+        </div>
       )
     }
 
@@ -126,7 +125,7 @@ class LoginComponent extends Component {
             this.iconRef = e
           }}
         >
-          {iconContent}
+          {buttonContent}
         </div>
       </ButtonWithIcon>
     )